add styling to toaster and quality menu settings
