+++ tubesize/src/styles/options.css
@@ -43,8 +43,7 @@ body {
 .description {
     color: #aaa;
     font-size: 13px;
-    margin-bottom: 12px;
-    text-align: center;
+    margin-bottom: 6px;
 }
 #options-grid {
     display: grid;
@@ -63,6 +62,21 @@ body {
     transition: all 0.2s;
 }
 .option-item:hover {
+    background: rgba(255, 255, 255, 0.08);
+    border-color: rgba(255, 255, 255, 0.15);
+}
+.setting-toggle-row {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 8px;
+    background: rgba(255, 255, 255, 0.04);
+    padding: 8px 10px;
+    border-radius: 6px;
+    border: 1px solid transparent;
+    transition: all 0.2s;
+}
+.setting-toggle-row:hover {
     background: rgba(255, 255, 255, 0.08);
     border-color: rgba(255, 255, 255, 0.15);
 }
